# Improve performance of array_find() etc