@TimWolla 2025-03-27
> I'm also wondering whether `Z_PARAM_FUNC()` may benefit from optimizations for `\Closure`.

It probably does. See PR description in: https://github.com/symfony/symfony/pull/60000.

I planned to have a look at the performance differences between `Closure` and `callable` as parameter types already.