@TimWolla 2025-03-28
For reference, performance numbers after the latest changes with the extra function:

```
Benchmark 1: /tmp/bench/php-before test.php
  Time (mean ± σ):     667.0 ms ±   3.7 ms    [User: 662.7 ms, System: 2.8 ms]
  Range (min … max):   663.1 ms … 674.3 ms    10 runs
 
Benchmark 2: /tmp/bench/php-after test.php
  Time (mean ± σ):     650.5 ms ±   3.3 ms    [User: 646.9 ms, System: 2.0 ms]
  Range (min … max):   647.1 ms … 657.7 ms    10 runs
 
Summary
  /tmp/bench/php-after test.php ran
    1.03 ± 0.01 times faster than /tmp/bench/php-before test.php
```

No idea why there even is a 70ms difference in the new baseline …