@nielsdos 2025-03-28
@TimWolla And if you add `zend_always_inline` to `php_is_true` ?