@TimWolla 2025-03-29
@nielsdos Doesn't make a difference. And also note how the `php-before` has gotten slower (with the commit being 370f24290f5d204ad3b1625b02a73d1fc64ea11a for the newer test). So it's probably my CPU being cursed again.